Deal properly with non-translated strings
By default, jall throws an exception when there's no translation for that
string.

An option can be used to change that behavior, in which case, jall returns the
string ID instead of throwing an exception.

This patch changes the previously described behaviour, by returning, instead,
the string in the defaultLanguage. Only if the string isn't available on the
defaultLanguage either, then the ID will be returned instead.

Closes: #6